Set up fzf as a vim plugin and make it use rg by default
I'm not entirely comfortable using this on the commandline yet (it
overrides some defaults and I don't want to set it up without a clear
usecase for myself), but I really like this tool in vim.

I overrode the default command that it uses to list files with ripgrep
because it's way faster and also respects files ignored with .gitignore.
Ripgrep can be configured to search ignored files with the flag
--no-ignore-vcs.