Update gitignore to whitelist instead of blacklist
My gitignore started out as the output of ls, and I deleted all the
files that I wanted to keep under source control. This tends to
accumulate a lot of cruft and reveals more about my system than is
necessary, not to mention that it requires a lot of upkeep every time I
install something new.