Save fewer things in vim sessionos
I really only want buffers and curdir, but I left a few things that
looked inoffensive. Mainly I removed `options` from the list because it
was doing weird things with colors for some reason. My vimrc should
already have run at the time we're loading a session, so it should have
handled setting all of the options that I actually want to be
persistent.