Add experimental tmux pane highlighting
This doesn't work for vim panes because my vim colorscheme is kind of
greedy about setting its colors. There was a quick hack suggested to
just comment out the `hi Normal` line that sets ctermbg and guibg, but
my colorscheme is trying to be fancy and share code by having one
function that can set colors for a highlight group passed in, so it's
nontrivial to quickly hack the Normal group without completely breaking
the colorscheme.

I'm also not sure if I want the pane bg colors to match the active
window or the inactive window -- currently they match the active but
I might switch that depending on how this looks on other monitors.